ulta/requirements.txt
python-dateutil
scrapy
